requirements.txt
flask==2.2.5
werkzeug==2.2.3
flask-cors==3.0.10
yt-dlp==2023.3.4
redis==6.2.0
rq==2.4.0
python-dotenv==1.1.0 